feat: modify printLotterries to print manual
