+++ src/main/java/lotto/model/Inventory.java
@@ -1,5 +1,7 @@
 package lotto.model;
 
+import lotto.controller.LotteryController;
+import lotto.view.InputView;
 import lotto.view.ResultView;
 
 import java.util.ArrayList;
@@ -7,10 +9,14 @@ import java.util.List;
 
 public class Inventory {
     private List<Lottery> lotteries;
+    private List<Lottery> manualLotteries;
 
-    public Inventory(){
+    public Inventory() {
         this.lotteries = new ArrayList();
-    };
+        this.manualLotteries = new ArrayList<>();
+    }
+
+    ;
 
     public Inventory(List<Lottery> lotteries) {
         this.lotteries = lotteries;
@@ -35,7 +41,8 @@ public class Inventory {
     }
 
     public void printLotteries() {
-        ResultView.printLotteries(this.lotteries);
+        ResultView.printLotteries(this.manualLotteries, "manual");
+        ResultView.printLotteries(this.lotteries, "auto");
     }
 
     public List<Reward> findWins(Winning winning) {
@@ -46,4 +53,20 @@ public class Inventory {
         }
         return rewards;
     }
+
+    public void scanManualLotteries(int amount) {
+        List<Lottery> manualLotteries = new ArrayList<>();
+        System.out.println("Put manual lottery.(" + amount + " times)");
+        for (int i = 0; i < amount; i++) {
+            manualLotteries.add(scanManualLottery());
+        }
+        this.manualLotteries = manualLotteries;
+    }
+
+    private Lottery scanManualLottery() {
+        String scanned = InputView.scan();
+        List<LotteryNumber> lotteryNumbers = LotteryController.parseNumbers(scanned);
+        return new Lottery(lotteryNumbers);
+    }
+
 }
